<message> Just noticed I didn't do pagination (more on the commit body...)
Despite I have an idea of what to do to create it, I started to feel lazy considering the
effort I just put in this. It's quite big already, since I was going for interfaces
interfaces everywhere and abstracting as much as possible.

Hope you don't have too much trouble reviewing it, but don't hesitate to ask.